Select the browser profile with --profile, and clear the lint backlog
Switching identity was a config edit, which is the wrong shape for something
changed per session and easy to leave switched. `scripts/akagi run --profile
west` now does it, and the script creates and links that profile before launch
so the first run of a new name lands in data/profiles/ rather than inside the
disposable target tree.

The override is held on AppState rather than merged into the loaded config.
The frontend reads that config and writes it back whenever any setting is
saved, so an override living there would be persisted by an unrelated edit and
a run-scoped identity would silently become permanent. Applied instead where
capture starts, and logged when it disagrees with the file.

Also clears the lints that had accumulated: two same-type casts, two needless
box allocations, a question-mark rewrite, a range loop over an indexed slice,
and the rustfmt drift across six files. Workspace clippy and fmt are both
clean now, so future noise will be attributable.

+++ native_bot/src/obs.rs
@@ -271,13 +271,13 @@ impl EncInput {
 
         // --- round wind (one-hot on honor position) ---
         // wind i lives at tile34 27+i; encode via tile id (27+i)*4.
-        if let Some(idx) = tile_index((27 + self.round_wind.min(3)) as u8 * 4, np) {
+        if let Some(idx) = tile_index((27 + self.round_wind.min(3)) * 4, np) {
             set!(ch, idx, 1.0);
         }
         ch += 1;
 
         // --- self seat wind (one-hot on honor position) ---
-        if let Some(idx) = tile_index((27 + self.seat_wind.min(3)) as u8 * 4, np) {
+        if let Some(idx) = tile_index((27 + self.seat_wind.min(3)) * 4, np) {
             set!(ch, idx, 1.0);
         }
         ch += 1;
